Create Replacing If Else Chains with Switch.js
FreeCodeCamp will not accept this answer, as it will detect the code in the multi-line comment as working code, so it will tell you that you did not replace the if-else statements. However, I put it in a multi-line comment here in the event that anyone might want to see what the if-else which was replaced is.